requirements.txt
asgiref==3.7.2
numpy==1.24.3
Pillow==9.5.0
PyYAML==6.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
